Flow: use cobra for the CLI and add an `agent run` command
